addeed support for large numbers

+++ inputformat/src/main/java/org/zuinnote/hadoop/bitcoin/format/common/BitcoinTransactionOutput.java
@@ -17,6 +17,7 @@
 package org.zuinnote.hadoop.bitcoin.format.common;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 
 public class BitcoinTransactionOutput implements Serializable {
 
@@ -25,17 +26,17 @@ public class BitcoinTransactionOutput implements Serializable {
 	 */
 	private static final long serialVersionUID = 2854570630540937753L;
 	
-private long value;
+private BigInteger value;
 private byte[] txOutScriptLength;
 private byte[] txOutScript;
 
-public BitcoinTransactionOutput(long value, byte[] txOutScriptLength, byte[] txOutScript) {
+public BitcoinTransactionOutput(BigInteger value, byte[] txOutScriptLength, byte[] txOutScript) {
 	this.value=value;
 	this.txOutScriptLength=txOutScriptLength;
 	this.txOutScript=txOutScript;
 }
 
-public long getValue() {
+public BigInteger getValue() {
 	return this.value;
 }
